Support the /metadata endpoint in DCOSClient
This endpoint returns the cluster's IP as well as the cluster ID.

This also moves the DC/OS client to its own package
